Refactor uploader-blobFS into the sender- structure
To be consistent with AzureFiles and Blob
